fix tests... again
